decomponse toolbar into smaller files
